pyproject.toml
[project]
name = "ii-lib-managment"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14,<4.0"
dependencies = [
    "uvicorn (>=0.41.0,<0.42.0)",
    "fastapi (>=0.135.1,<0.136.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "markitdown[pdf] (>=0.1.5,<0.2.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "lightrag-hku (>=1.4.10,<2.0.0)",
    "google-genai (>=1.66.0,<2.0.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "minio (>=7.2.20,<8.0.0)",
    "langgraph (>=1.0.10,<2.0.0)",
    "pydantic-ai (>=1.67.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
